feature: move log to util
